Merged from matamx/druid

+++ indexing-service/src/main/java/io/druid/indexing/overlord/scaling/EC2AutoScalingStrategy.java
@@ -125,8 +125,7 @@ public class EC2AutoScalingStrategy implements AutoScalingStrategy
                   return input.getInstanceId();
                 }
               }
-          ),
-          result.getReservation().getInstances()
+          )
       );
     }
     catch (Exception e) {
@@ -140,7 +139,7 @@ public class EC2AutoScalingStrategy implements AutoScalingStrategy
   public AutoScalingData terminate(List<String> ips)
   {
     if (ips.isEmpty()) {
-      return new AutoScalingData(Lists.<String>newArrayList(), Lists.<Instance>newArrayList());
+      return new AutoScalingData(Lists.<String>newArrayList());
     }
 
     DescribeInstancesResult result = amazonEC2Client.describeInstances(
@@ -156,37 +155,41 @@ public class EC2AutoScalingStrategy implements AutoScalingStrategy
     }
 
     try {
-      log.info("Terminating instance[%s]", instances);
-      amazonEC2Client.terminateInstances(
-          new TerminateInstancesRequest(
-              Lists.transform(
-                  instances,
-                  new Function<Instance, String>()
-                  {
-                    @Override
-                    public String apply(Instance input)
-                    {
-                      return input.getInstanceId();
-                    }
-                  }
-              )
-          )
-      );
-
-      return new AutoScalingData(
+      return terminateWithIds(
           Lists.transform(
-              ips,
-              new Function<String, String>()
+              instances,
+              new Function<Instance, String>()
               {
                 @Override
-                public String apply(@Nullable String input)
+                public String apply(Instance input)
                 {
-                  return String.format("%s:%s", input, config.getWorkerPort());
+                  return input.getInstanceId();
                 }
               }
-          ),
-          instances
+          )
       );
+    }
+    catch (Exception e) {
+      log.error(e, "Unable to terminate any instances.");
+    }
+
+    return null;
+  }
+
+  @Override
+  public AutoScalingData terminateWithIds(List<String> ids)
+  {
+    if (ids.isEmpty()) {
+      return new AutoScalingData(Lists.<String>newArrayList());
+    }
+
+    try {
+      log.info("Terminating instances[%s]", ids);
+      amazonEC2Client.terminateInstances(
+          new TerminateInstancesRequest(ids)
+      );
+
+      return new AutoScalingData(ids);
     }
     catch (Exception e) {
       log.error(e, "Unable to terminate any instances.");
